+++ source/renderer/app/containers/Root.js
@@ -12,8 +12,16 @@ type Props = InjectedContainerProps;
 export default class Root extends Component<Props> {
   render() {
     const { stores, actions, children } = this.props;
-    const { networkStatus, profile, adaRedemption, app, wallets } = stores;
-    const { isBlockConsolidationStatusDialog, isDelegationPage } = app;
+    const {
+      networkStatus,
+      profile,
+      adaRedemption,
+      wallets,
+      app,
+      staking,
+    } = stores;
+    const { isBlockConsolidationStatusDialog } = app;
+    const { isTakingPage } = staking;
     const { isProfilePage, isSettingsPage } = profile;
     const { isAdaRedemptionPage } = adaRedemption;
     const { hasLoadedWallets } = wallets;
@@ -27,7 +35,7 @@ export default class Root extends Component<Props> {
 
     const isPageThatDoesntNeedWallets =
       isBlockConsolidationStatusDialog ||
-      ((isAdaRedemptionPage || isDelegationPage || isSettingsPage) &&
+      ((isAdaRedemptionPage || isTakingPage || isSettingsPage) &&
         hasLoadedWallets &&
         isSynced);